Extract out most of the js/markdown shared logic
