+++ src/language-js/embed/js.js
@@ -4,18 +4,11 @@ const {
   builders: { indent, softline, literalline, dedentToRoot },
 } = require("../../document/index.js");
 const { escapeTemplateCharacters } = require("../print/template-literal.js");
+const { unescape, dedent } = require("./utils.js");
 
 function format(path, print, textToDoc, parser) {
   const node = path.getValue();
-  let text = node.quasis[0].value.raw.replace(
-    /((?:\\\\)*)\\`/g,
-    (_, backslashes) => "\\".repeat(backslashes.length / 2) + "`"
-  );
-  const indentation = getIndentation(text);
-  const hasIndent = indentation !== "";
-  if (hasIndent) {
-    text = text.replace(new RegExp(`^${indentation}`, "gm"), "");
-  }
+  const { text, hasIndent } = dedent(unescape(node.quasis[0].value.raw));
   const doc = escapeTemplateCharacters(
     textToDoc(text, { parser }, { stripTrailingHardline: true }),
     true
@@ -26,11 +19,6 @@ function format(path, print, textToDoc, parser) {
     softline,
     "`",
   ];
-}
-
-function getIndentation(str) {
-  const firstMatchedIndent = str.match(/^([^\S\n]*)\S/m);
-  return firstMatchedIndent === null ? "" : firstMatchedIndent[1];
 }
 
 module.exports = format;
